+++ app/src/main/java/org/stepik/android/exams/adaptive/model/Card.kt
@@ -11,7 +11,7 @@ import org.stepik.android.exams.api.StepicRestService
 import org.stepik.android.exams.data.preference.SharedPreferenceHelper
 import org.stepik.android.exams.di.qualifiers.BackgroundScheduler
 import org.stepik.android.exams.di.qualifiers.MainScheduler
-import org.stepik.android.exams.web.AttemptRequest
+import org.stepik.android.exams.data.model.AttemptRequest
 import org.stepik.android.model.Lesson
 import org.stepik.android.model.Step
 import org.stepik.android.model.attempts.Attempt
